requirements.txt
streamlit==1.42.2
pandas==2.2.3
numpy==1.26.4
tld==0.13
trafilatura==2.0.0
courlan==1.3.2
pymupdf==1.25.3
beautifulsoup4==4.13.3
requests==2.32.3
pdfminer.six==20221105
tzlocal==5.3
anthropic==0.3.11
google-generativeai==0.3.2
fpdf==1.7.2
python-docx==1.1.0


